rename markdown-generator.js to contributions-readme-generator.js * move readme generator to its own file
